Begin relicensing to AGPLv3
Because CycloMonitor is mainly interacted with over the internet, it is essential that all users of the bot have the ability to access the source code, regardless of whom is hosting the bot. Relicensing to the GNU AGPL v3 ensures that anyone that runs a modified copy of the bot is required to provide its users with its source code.